# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "server",
    "description": "",
    "license": {
      "name": "AGPL-3.0-only",
      "identifier": "AGPL-3.0-only"
    },
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/performances": {
      "get": {
        "tags": [
          "performances"
        ],
        "operationId": "list_performances",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number, 1-indexed. Defaults to 1.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "Items per page. Defaults to 20. The server enforces a maximum.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "Text search across performance title, song title, and singer names.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "song_id",
            "in": "query",
            "description": "Filter by song ID. Returns only performances that include this song.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Field to sort by. Defaults to `performance_date`.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/PerformanceSort"
            }
          },
          {
            "name": "sort_dir",
            "in": "query",
            "description": "Sort direction. Defaults to `desc`.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortDir"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paged list of performances",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponse_PerformanceSummary"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "performances"
        ],
        "operationId": "create_performance",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePerformanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created performance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerformanceResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/performances/{id}": {
      "get": {
        "tags": [
          "performances"
        ],
        "operationId": "get_performance",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Performance ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Performance detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerformanceResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "performances"
        ],
        "operationId": "update_performance",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Performance ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePerformanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated performance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerformanceResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "delete": {
        "tags": [
          "performances"
        ],
        "operationId": "delete_performance",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Performance ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/performances/{id}/audio": {
      "post": {
        "tags": [
          "performances"
        ],
        "operationId": "upload_audio",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Performance ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/AudioUpload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Audio uploaded and linked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioInfo"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Performance not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/performances/{id}/audio/link": {
      "post": {
        "tags": [
          "performances"
        ],
        "operationId": "add_audio_link",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Performance ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddAudioLinkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Audio link created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioInfo"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Performance not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/performances/{id}/audio/{asset_id}": {
      "delete": {
        "tags": [
          "performances"
        ],
        "operationId": "delete_audio",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Performance ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "asset_id",
            "in": "path",
            "description": "Audio asset ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Unlinked"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "patch": {
        "tags": [
          "performances"
        ],
        "operationId": "update_audio_kind",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Performance ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "asset_id",
            "in": "path",
            "description": "Audio asset ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAudioKindRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Kind updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioInfo"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatePerformanceRequest": {
        "type": "object",
        "description": "Request body for `POST /api/performances`.",
        "required": [
          "performance_date",
          "stream_number",
          "performance_number",
          "song_ids",
          "singer_ids",
          "tags"
        ],
        "properties": {
          "duration": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Duration in seconds.",
            "minimum": 0
          },
          "lyrics": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional inline lyrics content. Creates a lyrics row in a single round trip."
          },
          "performance_date": {
            "type": "string",
            "format": "date",
            "description": "Calendar date of the stream in which this performance occurred."
          },
          "performance_number": {
            "type": "integer",
            "format": "int32",
            "description": "Performance index within `performance_date` + `stream_number`, 1-based.",
            "minimum": 0
          },
          "singer_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "IDs of artists who sang in this performance."
          },
          "song_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "IDs of songs featured in this performance."
          },
          "stream_number": {
            "type": "integer",
            "format": "int32",
            "description": "Stream index within `performance_date`, 1-based.",
            "minimum": 0
          },
          "stream_time": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Offset in seconds from the start of the stream.",
            "minimum": 0
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PerformanceTagAssignment"
            },
            "description": "Tags to apply at creation time."
          },
          "title": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional display title override."
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Body returned for all error responses.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human readable description of the error."
          }
        }
      },
...
      "PerformanceSort": {
        "type": "string",
        "description": "Field to sort performances by in list endpoints.",
        "enum": [
          "performance_date",
          "play_count",
          "duration"
        ]
      },
...
      "SortDir": {
        "type": "string",
        "description": "Sort direction for list endpoints.",
        "enum": [
          "asc",
          "desc"
        ]
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    }
  }
}